moved dag utils module
