+++ dags/grobid_train_evaluate_source_dataset.py
@@ -16,7 +16,7 @@ from sciencebeam_airflow.dags.dag_conf import (
     get_eval_output_path
 )
 
-from sciencebeam_dag_utils import (
+from sciencebeam_airflow.dags.utils import (
     get_default_args,
     create_trigger_operator,
     create_trigger_next_task_dag_operator,
