+++ dags/sciencebeam_autocut_convert_training_data.py
@@ -4,7 +4,7 @@ from airflow.models import DAG
 
 from sciencebeam_airflow.dags.dag_ids import ScienceBeamDagIds
 
-from sciencebeam_dag_utils import (
+from sciencebeam_airflow.dags.utils import (
     get_default_args
 )
