+++ dags/sciencebeam_evaluate.py
@@ -8,7 +8,7 @@ from sciencebeam_airflow.utils.container_operators import ContainerRunOperator
 
 from sciencebeam_airflow.dags.dag_ids import ScienceBeamDagIds
 
-from sciencebeam_dag_utils import (
+from sciencebeam_airflow.dags.utils import (
     get_default_args,
     create_validate_config_operation,
     create_trigger_next_task_dag_operator,
