+++ dags/sciencebeam_evaluation_results_to_bq.py
@@ -12,7 +12,7 @@ from airflow.models import DAG, DagRun
 
 from sciencebeam_airflow.dags.dag_ids import ScienceBeamDagIds
 
-from sciencebeam_dag_utils import (
+from sciencebeam_airflow.dags.utils import (
     get_default_args,
     create_validate_config_operation,
     create_trigger_next_task_dag_operator,
